+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Supported rust types are `Vec<char>`, `Vec<u32>`, and `String`
   (and, of course, `[char]`, `[u32]`, and `str` for serialization).
 - Added support for the `a` type, which is just an alias of `S`.
+- Added the `arrayvec` feature.  This enables deserialization of
+  type `U` without individual allocations per element.
 
 ### Changed
 - `np.datetime64` now uses i64 instead of u64 for serialization, as it is